bracket-pathfinding: Disable threaded feature on wasm32 target.
Cargo doesn't have a good way of allowing features to be enabled on a
per-platform basis. This makes selecting the threaded feature all-or-nothing:
either it's enabled and you can't build for wasm, or it's disabled, even for
platforms that support threading.

Fix this by making the feature a no-op on platforms that don't support it.